Refactor fss monitor
Calculate used disk space with java code instead of using du command

LMCROSSITXSADEPLOY-479

+++ com.sap.cloud.lm.sl.cf.web/src/main/java/com/sap/cloud/lm/sl/cf/web/monitoring/MetricsMBean.java
@@ -2,8 +2,8 @@ package com.sap.cloud.lm.sl.cf.web.monitoring;
 
 public interface MetricsMBean {
 
-    double getUsedFssSpace();
+    long getUsedFssSpace();
 
-    double getUsedContainerSpace();
+    long getUsedContainerSpace();
 
 }
